update excel structure

+++ src/main/java/com/duyhung/lydinc_backend/model/dto/UniversityDto.java
@@ -6,19 +6,15 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.util.List;
-
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 public class UniversityDto {
-
     private Integer universityId;
     private String shortName;
     private String fullName;
     private String logo;
     private String location;
-    private int studentCount;
-
+    private Long studentCount;
 }
